change ilbSSL to check if there a certificate already

+++ resources/uk/gov/hmcts/contino/ilbSSL.sh
@@ -32,21 +32,34 @@ EOF
 commonName=core-compute-$platform
 sed -i "s/commonNameVar/$commonName/g" $domain.conf
 
-openssl req -new -sha256 -nodes -out $domain.csr -newkey rsa:2048 -keyout $domain.key -config <( cat $domain.conf )
 
-openssl x509 -req -in $domain.csr -signkey $domain.key -out $domain.cer
+certjson=$(env AZURE_CONFIG_DIR=/opt/jenkins/.azure-$subscription az keyvault certificate show --vault-name app-vault-${subscription} -n $domain)
+if [[ $certjson = *"not found"* ]]; then
+  echo "Cert doesn't exist... will create one"
+  openssl req -new -sha256 -nodes -out $domain.csr -newkey rsa:2048 -keyout $domain.key -config <( cat $domain.conf )
 
-openssl pkcs12 -export -in $domain.cer -inkey $domain.key -out $domain.pfx -password pass:$pfxPass
+  openssl x509 -req -in $domain.csr -signkey $domain.key -out $domain.cer
 
-rm -f $domain.key $domain.csr $domain.conf
+  openssl pkcs12 -export -in $domain.cer -inkey $domain.key -out $domain.pfx -password pass:$pfxPass
 
-env AZURE_CONFIG_DIR=/opt/jenkins/.azure-$subscription az keyvault certificate import --vault-name app-vault-${subscription} -n $domain -f $domain.pfx --password $pfxPass
+  rm -f $domain.key $domain.csr $domain.conf
+
+  env AZURE_CONFIG_DIR=/opt/jenkins/.azure-$subscription az keyvault certificate import --vault-name app-vault-${subscription} -n $domain -f $domain.pfx --password $pfxPass
+
+  # get base64 representation of pfx file and write to file
+  a=$(cat ./$domain.pfx)
+  echo -n $a | base64 > base64.txt
+  truncate -s -1 base64.txt
+else
+  echo "Cert found... Pulling pfxBlobString for existing one"
+  env AZURE_CONFIG_DIR=/opt/jenkins/.azure-$subscription az keyvault certificate show --vault-name app-vault-${subscription} --name $domain --query cer -o tsv > $domain.cer
+  env AZURE_CONFIG_DIR=/opt/jenkins/.azure-$subscription az keyvault certificate show --vault-name app-vault-${subscription} --name $domain --query x509ThumbprintHex -o tsv > base64.txt
+fi
 
 #to be put back in
+echo "Pushing cert to app gateway..."
 env AZURE_CONFIG_DIR=/opt/jenkins/.azure-$subscription az network application-gateway auth-cert create --cert-file $domain.cer --gateway-name $domain --name $domain --resource-group core-infra-${platform}
+rm -f $domain.cer
 
-# get base64 representation of pfx file and write to file
-a=$(cat ./$domain.pfx)
-echo -n $a | base64 > base64.txt
-truncate -s -1 base64.txt
+
